Conflicts:
	components/tools/OmeroWeb/omeroweb/webgateway/views.py

+++ components/tools/OmeroPy/src/omero/gateway/__init__.py
@@ -25,7 +25,8 @@ import ConfigParser
 import omero
 import omero.clients
 from omero.util.decorators import timeit, TimeIt, setsessiongroup
-from omero.cmd import Chgrp
+from omero.cmd import Chgrp, DoAll
+from omero.api import Save
 from omero.callbacks import CmdCallbackI
 from omero.gateway.utils import ServiceOptsDict, GatewayConfig
 import omero.scripts as scripts
@@ -396,17 +397,6 @@ class BlitzObjectWrapper (object):
         """
         return self._conn.canOwnerWrite(self)
     
-    def canDelete(self):
-        """
-        Determines whether the current user can delete this object.
-        Returns True if the object L{isOwned} by the current user or L{isLeaded}
-        (current user is leader of this the group that this object belongs to)
-        
-        @rtype:     Boolean
-        @return:    see above
-        """
-        return self.isOwned() or self.isLeaded()
-    
     def isOwned(self):
         """
         Returns True if the object owner is the same user specified in the connection's Event Context
@@ -414,7 +404,7 @@ class BlitzObjectWrapper (object):
         @rtype:     Boolean
         @return:    True if current user owns this object
         """
-        return (self._obj.details.owner.id.val == self._conn.getEventContext().userId)
+        return (self._obj.details.owner.id.val == self._conn.getUserId())
     
     def isLeaded(self):
         """
@@ -701,11 +691,14 @@ class BlitzObjectWrapper (object):
         """ Loads the annotation links for the object (if not already loaded) and saves them to the object """
         if not hasattr(self._obj, 'isAnnotationLinksLoaded'): #pragma: no cover
             raise NotImplementedError
+        # Need to set group context. If '-1' then canDelete() etc on annotations will be False
+        ctx = self._conn.SERVICE_OPTS.copy()
+        ctx.setOmeroGroup(self.details.group.id.val)
         if not self._obj.isAnnotationLinksLoaded():
             query = "select l from %sAnnotationLink as l join fetch l.details.owner join fetch l.details.creationEvent "\
             "join fetch l.child as a join fetch a.details.owner join fetch a.details.creationEvent "\
             "where l.parent.id=%i" % (self.OMERO_CLASS, self._oid)
-            links = self._conn.getQueryService().findAllByQuery(query, None, self._conn.SERVICE_OPTS)
+            links = self._conn.getQueryService().findAllByQuery(query, None, ctx)
             self._obj._annotationLinksLoaded = True
             self._obj._annotationLinksSeq = links
 
@@ -742,12 +735,15 @@ class BlitzObjectWrapper (object):
 
         # Using omero.cmd.Delete rather than deleteObjects since we need
         # spec/id pairs rather than spec+id_list as arguments
-        handle = self._conn.c.sf.submit(dcs, self._conn.SERVICE_OPTS)
-        callback = omero.callbacks.CmdCallbackI(self._conn.c, handle)
-        # Maximum wait time 5 seconds, will raise a LockTimeout if the
-        # delete has not finished by then.
-        callback.loop(10, 500)
-        self._obj.unloadAnnotationLinks()
+        if len(dcs):
+            doall = omero.cmd.DoAll()
+            doall.requests = dcs
+            handle = self._conn.c.sf.submit(doall, self._conn.SERVICE_OPTS)
+            try:
+                self._conn._waitOnCmd(handle)
+            finally:
+                handle.close()
+            self._obj.unloadAnnotationLinks()
 
     def removeAnnotations (self, ns):
         """
@@ -763,12 +759,13 @@ class BlitzObjectWrapper (object):
         for al in self._getAnnotationLinks(ns=ns):
             a = al.child
             ids.append(a.id.val)
-        handle = self._conn.deleteObjects('/Annotation', ids)
-        callback = omero.callbacks.CmdCallbackI(self._conn.c, handle)
-        # Maximum wait time 5 seconds, will raise a LockTimeout if the
-        # delete has not finished by then.
-        callback.loop(10, 500)
-        self._obj.unloadAnnotationLinks()        
+        if len(ids):
+            handle = self._conn.deleteObjects('/Annotation', ids)
+            try:
+                self._conn._waitOnCmd(handle)
+            finally:
+                handle.close()
+            self._obj.unloadAnnotationLinks()        
     
     # findAnnotations(self, ns=[])
     def getAnnotation (self, ns=None):
@@ -895,7 +892,7 @@ class BlitzObjectWrapper (object):
         if sameOwner:
             d = self.getDetails()
             ad = ann.getDetails()
-            if self._conn.isAdmin() and self._conn._userid != d.getOwner().id:
+            if self._conn.isAdmin() and self._conn.getUserId() != d.getOwner().id:
                 # Keep the annotation owner the same as the linked of object's
                 if ad.getOwner() and d.getOwner().omeName == ad.getOwner().omeName and d.getGroup().name == ad.getGroup().name:
                     newConn = ann._conn
@@ -1414,7 +1411,7 @@ class _BlitzGateway (object):
             if self.c.sf is None: #pragma: no cover
                 logger.debug('... c.sf is None, reconnecting')
                 return self.connect()
-            return self.c.sf.keepAlive(self._proxies['admin']._obj)
+            return self.c.sf.keepAlive(self._proxies['admin']._getObj())
         except Ice.ObjectNotExistException: #pragma: no cover
             # The connection is there, but it has been reset, because the proxy no longer exists...
             logger.debug(traceback.format_exc())
@@ -1523,14 +1520,9 @@ class _BlitzGateway (object):
             self._proxies['timeline'] = ProxyObjectWrapper(self, 'getTimelineService')
             self._proxies['types'] = ProxyObjectWrapper(self, 'getTypesService')
             self._proxies['update'] = ProxyObjectWrapper(self, 'getUpdateService')
-        self._ctx = self._proxies['admin'].getEventContext()
-        if self._ctx is not None:
-            self._userid = self._ctx.userId
-            self._username = self._ctx.userName
-        else:
-            self._userid = None
-            self._username = None
-            self._user = None
+        self._userid = None
+        self._user = None
+        self._ctx = None
 
         if self._session_cb: #pragma: no cover
             if self._was_join:
@@ -1651,7 +1643,7 @@ class _BlitzGateway (object):
                     s = self.c.joinSession(self._sessionUuid)   # timeout to allow this is $ omero config set omero.sessions.timeout 3600000
                     s.detachOnDestroy()
                     self.SERVICE_OPTS = self.createServiceOptsDict()
-                    logger.debug('Joined Session OK with Uuid: %s and timeToIdle: %s, timeToLive: %s' % (self._sessionUuid, self.getSession().timeToIdle.val, self.getSession().timeToLive.val))
+                    logger.debug('Joined Session OK with Uuid: %s' % (self._sessionUuid,))
                     self._was_join = True
                 except Ice.SyscallException: #pragma: no cover
                     raise
@@ -1683,7 +1675,7 @@ class _BlitzGateway (object):
                 try:
                     logger.debug("Creating Session...")
                     self._createSession()
-                    logger.debug("Session created with timeout: %s & timeToLive: %s" % (self.getSession().timeToIdle.val, self.getSession().timeToLive.val))
+                    logger.debug("Session created")
                 except omero.SecurityViolation:
                     if self.group is not None:
                         # User don't have access to group
@@ -1772,6 +1764,24 @@ class _BlitzGateway (object):
             self._ctx = self._proxies['admin'].getEventContext()
         return self._ctx
 
+    def getUserId (self):
+        """
+        Returns current experimenter id
+
+        @return:    Current Experimenter id
+        @rtype:     long
+        """
+        if self._userid is None:
+            self._userid = self.getEventContext().userId
+        return self._userid
+
+    def setUserId (self, uid):
+        """
+        Sets current experimenter id
+        """
+        self._userid = uid
+        self._user = None
+            
     def getUser (self):
         """
         Returns current Experimenter.
@@ -1779,10 +1789,10 @@ class _BlitzGateway (object):
         @return:    Current Experimenter
         @rtype:     L{ExperimenterWrapper}
         """
-        if self._ctx is None:
-            return None
         if self._user is None:
-            self._user = self._ctx.userName!="guest" and self.getObject("Experimenter", self._userid) or None
+            uid = self.getUserId()
+            if uid is not None:
+                self._user = self.getObject("Experimenter", self._userid) or None
         return self._user
     
     def getGroupFromContext(self):
@@ -1835,7 +1845,7 @@ class _BlitzGateway (object):
         @return:    Boolean
         """
         
-        return self.isAdmin() or (self._userid == obj.getDetails().getOwner().getId() and
+        return self.isAdmin() or (self.getUserId() == obj.getDetails().getOwner().getId() and
                                   obj.getDetails().getPermissions().isUserWrite())
 
     def canOwnerWrite (self, obj):
@@ -2346,7 +2356,7 @@ class _BlitzGateway (object):
         default = self.getObject("ExperimenterGroup", self.getEventContext().groupId)
         if not default.isPrivate() or self.isLeader():
             for d in default.copyGroupExperimenterMap():
-                if d.child.id.val != self.getEventContext().userId:
+                if d.child.id.val != self.getUserId():
                     yield ExperimenterWrapper(self, d.child)
 
     def groupSummary(self, gid=None, exclude_self=False):
@@ -2362,7 +2372,7 @@ class _BlitzGateway (object):
             gid = self.getEventContext().groupId
         userId = None
         if exclude_self:
-            userId = self.getEventContext().userId
+            userId = self.getUserId()
         colleagues = []
         leaders = []
         default = self.getObject("ExperimenterGroup", gid)
@@ -2397,7 +2407,7 @@ class _BlitzGateway (object):
                 "exists ( select gem from GroupExperimenterMap as gem where gem.child = e.id " \
                 "and gem.parent.id in (:gids)) order by e.omeName"
         for e in q.findAllByQuery(sql, p,self.SERVICE_OPTS):
-            if e.id.val != self.getEventContext().userId:
+            if e.id.val != self.getUserId():
                 yield ExperimenterWrapper(self, e)
 
     def listOwnedGroups(self):
@@ -3152,16 +3162,16 @@ class _BlitzGateway (object):
         return prx
 
 
-    def chgrpObject(self, graph_spec, obj_id, group_id):
+    def chgrpObjects(self, graph_spec, obj_ids, group_id, container_id=None):
         """
-        Change the Group for a specified object using queue.
+        Change the Group for a specified objects using queue.
 
         @param graph_spec:      String to indicate the object type or graph
                                 specification. Examples include:
                                  * '/Image'
                                  * '/Project'   # will move contents too.
                                  * NB: Also supports 'Image' etc for convenience
-        @param obj_id:          ID for the object to move.
+        @param obj_ids:         IDs for the objects to move.
         @param group_id:        The group to move the data to.
         """
 
@@ -3169,9 +3179,31 @@ class _BlitzGateway (object):
             graph_spec = '/%s' % graph_spec
             logger.debug('chgrp Received object type, using "%s"' % graph_spec)
 
-        chgrp = omero.cmd.Chgrp(type=graph_spec, id=obj_id, options=None, grp=group_id)
+        # (link, child, parent)
+        parentLinkClasses = {"/Image": (omero.model.DatasetImageLinkI, omero.model.ImageI, omero.model.DatasetI),
+                        "/Dataset": (omero.model.ProjectDatasetLinkI, omero.model.DatasetI, omero.model.ProjectI),
+                        "/Plate": (omero.model.ScreenPlateLinkI, omero.model.PlateI, omero.model.ScreenI)}
+        da = DoAll()
+        requests = []
+        for obj_id in obj_ids:
+            obj_id = long(obj_id)
+            logger.debug('DoAll Chgrp: type: %s, id: %s, grp: %s' % (graph_spec, obj_id, group_id))
+            chgrp = omero.cmd.Chgrp(type=graph_spec, id=obj_id, options=None, grp=group_id)
+            requests.append(chgrp)
+            if container_id is not None and graph_spec in parentLinkClasses:
+                # get link class for graph_spec objects
+                link_klass = parentLinkClasses[graph_spec][0]
+                link = link_klass()
+                link.child = parentLinkClasses[graph_spec][1](obj_id, False)
+                link.parent = parentLinkClasses[graph_spec][2](container_id, False)
+                save = Save()
+                save.obj = link
+                requests.append(save)
 
-        prx = self.c.sf.submit(chgrp)
+        da.requests = requests
+        ctx = self.SERVICE_OPTS.copy()
+        ctx.setOmeroGroup(group_id)         # NB: For Save to work, we need to be in target group
+        prx = self.c.sf.submit(da, ctx)
         return prx
 
 
@@ -4306,7 +4338,7 @@ class _ExperimenterWrapper (BlitzObjectWrapper):
 
     def is_self(self):
         """ Returns True if this Experimenter is the current user """
-        return self.getId() == self._conn.getEventContext().userId
+        return self.getId() == self._conn.getUserId()
     
 ExperimenterWrapper = _ExperimenterWrapper
 
@@ -6033,6 +6065,13 @@ class _ImageWrapper (BlitzObjectWrapper):
             return (0, pmax-1)
 
     @assert_pixels
+    def requiresPixelsPyramid (self):
+        pixels_id = self._obj.getPrimaryPixels().getId().val
+        rp = self._conn.createRawPixelsStore()
+        rp.setPixelsId(pixels_id, True, self._conn.SERVICE_OPTS)
+        return rp.requiresPixelsPyramid()
+
+    @assert_pixels
     def getPrimaryPixels (self):
         """
         Loads pixels and returns object in a L{PixelsWrapper}
@@ -6068,16 +6107,19 @@ class _ImageWrapper (BlitzObjectWrapper):
         @type windows:      List of tuples. [(20, 300), (None, None), (50, 500)]. Must be tuples for all channels
         @param colors:      List of colors. ['F00', None, '00FF00'].  Must be item for each channel
         """
-
+        abs_channels = [abs(c) for c in channels]
+        idx = 0     # index of windows/colors args above
         for c in range(len(self.getChannels())):
             self._re.setActive(c, (c+1) in channels, self._conn.SERVICE_OPTS)
             if (c+1) in channels:
-                if windows is not None and windows[c][0] is not None and windows[c][1] is not None:
-                    self._re.setChannelWindow(c, *(windows[c] + [self._conn.SERVICE_OPTS]))
-                if colors is not None and colors[c]:
-                    rgba = splitHTMLColor(colors[c])
+                if windows is not None and windows[idx][0] is not None and windows[idx][1] is not None:
+                    self._re.setChannelWindow(c, *(windows[idx] + [self._conn.SERVICE_OPTS]))
+                if colors is not None and colors[idx]:
+                    rgba = splitHTMLColor(colors[idx])
                     if rgba:
                         self._re.setRGBA(c, *(rgba + [self._conn.SERVICE_OPTS]))
+            if (c+1 in abs_channels):
+                idx += 1
         return True
 
     def getProjections (self):
@@ -6371,19 +6413,19 @@ class _ImageWrapper (BlitzObjectWrapper):
 
 
     @assert_re()
-    def renderJpeg (self, z, t, compression=0.9):
+    def renderJpeg (self, z=None, t=None, compression=0.9):
         """
         Return the data from rendering image, compressed (and projected).
         Projection (or not) is specified by calling L{setProjection} before renderJpeg.
         
-        @param z:               The Z index. Ignored if projecting image. 
-        @param t:               The T index. 
+        @param z:               The Z index. Ignored if projecting image. If None, use defaultZ
+        @param t:               The T index. If None, use defaultT
         @param compression:     Compression level for jpeg
         @type compression:      Float
         """
         
-        self._pd.z = long(z)
-        self._pd.t = long(t)
+        self._pd.z = z is not None and long(z) or self._re.getDefaultZ()
+        self._pd.t = t is not None and long(t) or self._re.getDefaultT()
         try:
             if compression is not None:
                 try:
@@ -6596,7 +6638,12 @@ class _ImageWrapper (BlitzObjectWrapper):
         for chunk in ofw.getFileInChunks():
             outfile.write(chunk)
         outfile.close()
-        self._conn.deleteObjects('/OriginalFile', todel) # No error handling?
+        handle = self._conn.deleteObjects('/OriginalFile', todel)
+        try:
+            self._conn._waitOnCmd(handle)
+        finally:
+            handle.close()
+
         return os.path.splitext(f.name.val)[-1], f.mimetype.val
         
     def renderImage (self, z, t, compression=0.9):
@@ -7016,7 +7063,10 @@ class _ImageWrapper (BlitzObjectWrapper):
 
     def _deleteSettings(self):
         handle = self._conn.deleteObjects("/Image/Pixels/RenderingDef", [self.getId()])
-        self._conn._waitOnCmd(handle)
+        try:
+            self._conn._waitOnCmd(handle)
+        finally:
+            handle.close()
 
     def _collectRenderOptions (self):
         """
@@ -7153,7 +7203,7 @@ class _ImageWrapper (BlitzObjectWrapper):
 
         roiOptions = omero.api.RoiOptions()
         if eid:
-            roiOptions.userId = omero.rtypes.rlong(self._conn._userid)
+            roiOptions.userId = omero.rtypes.rlong(self._conn.getUserId())
         
         result = self._conn.getRoiService().findByImage(self.id, roiOptions)
         count = sum(1 for roi in result.rois if isValidROI(roi))
